+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-02-25
+
 ### Added
 
 -   Add limited one-way compatibility for public pins sharing from modded clients to vanilla non-modded clients.
@@ -30,7 +32,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Initial release.
 
-[Unreleased]: https://github.com/nbusseneau/BetterCartographyTable/compare/0.1.1...HEAD
+[Unreleased]: https://github.com/nbusseneau/BetterCartographyTable/compare/0.2.0...HEAD
+
+[0.2.0]: https://github.com/nbusseneau/BetterCartographyTable/compare/0.1.1...0.2.0
 
 [0.1.1]: https://github.com/nbusseneau/BetterCartographyTable/compare/0.1.0...0.1.1